refactor(combo): move the Anchor connection to core/ComboAnchorNet
The socket, receive thread and roster move out of ComboShip.cpp. sActiveGame and
sResyncPending stay private to the module; the three callers that reached into them
now go through ActiveGame() and TakeResyncPending().

SDL_net moves with it - the launcher no longer includes SDL at all.
